Rename file type to enum in test folder
