+++ __test__/jsonTemplate/enum.ts
@@ -1,0 +1,6 @@
+export enum TemplateName {
+  mainTranslationTemplateWithContext = 'mainTranslationTemplateWithContext.json',
+  mainTranslationTemplateWithText = 'mainTranslationTemplateWithText.json',
+  homeScreenTemplateWithContext = 'homeScreenTemplateWithContext.json',
+  homeScreenTemplateWithText = 'homeScreenTemplateWithText.json',
+}
